Feat/ 소켓io disconnect 함수 변경, friendly game process #104

+++ backend/nest_project/src/game/ladderGame.gateway.ts
@@ -3,9 +3,9 @@ import { Injectable, Logger, OnModuleInit } from '@nestjs/common';
 import { GameField, Ball, Paddle } from './interface/game.interface';
 import { ConnectedPlayerService } from './service/connectedPlayer.service';
 
-import { from, Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
-import { ConnectedSocket, MessageBody, OnGatewayConnection, OnGatewayDisconnect, SubscribeMessage, WebSocketGateway, WebSocketServer, WsResponse } from '@nestjs/websockets';
+// import { from, Observable } from 'rxjs';
+// import { map } from 'rxjs/operators';
+import { ConnectedSocket, MessageBody, OnGatewayConnection, OnGatewayDisconnect, SubscribeMessage, WebSocketGateway, WebSocketServer } from '@nestjs/websockets';
 import { Socket, Server } from 'socket.io';
 
 import * as jwt from 'jsonwebtoken';
@@ -15,9 +15,12 @@ import { MatchEntity } from './entities/match.entity';
 
 import { MatchService } from './service/match.service';
 import { HistoryService } from './service/history.service';
-import { HistoryEntity } from './entities/history.entity';
+// import { HistoryEntity } from './entities/history.entity';
 
 import { Player } from './dto/player.dto';
+import { match } from 'assert';
+import { dir } from 'console';
+// import { Player } from './interface/game.interface';
 
 
 let ladderQueue: Player[];
@@ -25,8 +28,14 @@ let resetQueueTime: number;
 let currentQueueTime: number;
 let ladderRange: number;
 
+// let gameFieldArr: GameField[];
+// const gameFieldMap = new Map();
+
+
 // todo: ladder_game, friendly_game 이외의 네임스페이스 처리하는 코드 필요
-// todo: 게임 시작하기 전에 대기 화면 -> test
+
+// todo: 삭제
+const logger2 = new Logger('LadderGameGateway');
 
 @Injectable()
 @WebSocketGateway({ namespace: 'ladder_game' })
@@ -47,40 +56,50 @@ export class LadderGameGateway implements OnGatewayConnection, OnGatewayDisconne
 		ladderQueue = [];
 		resetQueueTime = Date.now();
 		currentQueueTime = Date.now();
-		ladderRange = 500;
+		ladderRange = 100;
 		this.gameFieldArr = [];
 
-		setInterval(this.queueProcess, 1000);
+		setInterval(() => this.queueProcess(), 1000);
 	}
+
+
+	@WebSocketServer()
+	server: Server;
 
 
 	async onModuleInit() 
 	{
 		this.connectedPlayerService.deleteAll();
+		this.matchService.deleteAll();
 	}
 
 	async handleConnection(socket: Socket)
 	{
 		this.logger.log(`Ladder Game Server: socketId [ ${socket.id} ] connected.`);
-		// 토큰, User 데이터와 소켓 아이디 결합하여 Player 객체에 저장
-		// user의 소켓 id 정보
-		// 인증 관련 부분(토큰 및 user 정보 socket에 주입 )
 		const token = socket.handshake.headers.authorization;
-		// jiwolee님이 알려주신 대로 프론트에 추가 -> ok
+
 		
-		// //userId가 없는 경우 or userProfile이나 userEntity가 없는 경우 소켓 연결끊음
+		// //userId가 없는 경우 or userProfile이나 userEntity가 없는 경우 소켓 연결 끊음
 		const userId = jwt.decode(token.split('Bearer ')[1])['userId'];
 		if (!userId)
 		{
 			return socket.disconnect();
+			// 클라이언트에서 이벤트 인지하면 menu로 리다이렉션
 		}
+		socket.data.userId = userId;
 		const userProfile = await this.profileService.getUserProfileById(userId);
 		if (!userProfile)
 		{
 			return socket.disconnect();
+			// 클라이언트에서 이벤트 인지하면 menu로 리다이렉션
 		}
 
-		const current  = await this.connectedPlayerService.createPlayer(userId, socket.id);
+		const current = await this.connectedPlayerService.createPlayer(userId, socket.id);
+		if(!current)
+		{
+			return socket.disconnect();
+		}
+		this.profileService.ingame(userId);
 		ladderQueue.push(current);
 		this.logger.log(`current Player : ${current.id}, ${current.socketId}`);
 	}
@@ -89,60 +108,101 @@ export class LadderGameGateway implements OnGatewayConnection, OnGatewayDisconne
 	{
 		// socket.emit('Error', new UnauthorizedException());
 
-		// 게임 도중 끊긴 연결이면, 게임 종료(매치 패배 처리)
-		const player = await this.connectedPlayerService.getPlayerBySocketId(socket.id);
-		const match_id = (await this.matchService.getByPlayerId(player.id)).match_id;
-		if (match_id)
+		if (await this.connectedPlayerService.getPlayerBySocketId(socket.id))
 		{
-			await this.endGame(match_id, player.id);
-		}
-
-		// 큐 잡는 도중 끊긴 연결이면, 래더 큐 배열에서도 삭제
-		for (let i = 0; i < ladderQueue.length; ++i)
-		{
-			if (player.id === ladderQueue[i].id)
+			const player_id = (await this.connectedPlayerService.getPlayerBySocketId(socket.id)).id;
+			// 큐 잡는 도중 끊긴 연결이면, 래더 큐 배열에서도 삭제
+			for (let i = 0; i < ladderQueue.length; ++i)
 			{
-				ladderQueue.splice(i, 1);
-				break ;
+				if (player_id === ladderQueue[i].id)
+				{
+					ladderQueue.splice(i, 1);
+					break ;
+				}
+			}
+
+			if (await this.matchService.getByPlayerId(player_id) === null)
+			{
+				await this.connectedPlayerService.deletePlayer(player_id);
+				this.profileService.logOn(socket.data.userId);
+				socket.disconnect();
+				return ;
+			}
+			else
+			{
+				const match_id = (await this.matchService.getByPlayerId(player_id)).match_id;
+				// this.logger.log(`handleDIsconnect : match id ${match_id}`);
+				if (match_id)
+				{
+					await this.endGame(match_id, player_id);
+				}
+				await this.connectedPlayerService.deletePlayer(player_id);
+				this.profileService.logOn(socket.data.userId);
+				socket.disconnect();
+				return ;
 			}
 		}
 
-		this.endPlayer(socket.id, player.id);
+		if ((await this.connectedPlayerService.getPlayerBySocketId(socket.id)) === null)
+		{
+			this.profileService.logOn(socket.data.userId);
+			socket.disconnect();
+			return ;
+		}
+
+		// const player_id = (await this.connectedPlayerService.getPlayerBySocketId(socket.id)).id;
+		// if (await this.matchService.getByPlayerId(player_id))
+		// {
+		// 	const match_id = (await this.matchService.getByPlayerId(player_id)).match_id;
+		// 	// this.logger.log(`handleDIsconnect : match id ${match_id}`);
+		// 	if (match_id)
+		// 	{
+		// 		await this.endGame(match_id, player_id);
+		// 	}
+		// }
+		// this.logger.log(`handleDIsconnect : socket id ${socket.id}, player id ${player_id}`);
+
+		// await this.connectedPlayerService.deletePlayer(player_id);
+		// this.profileService.logOn(socket.data.userId);
+
 	}
 
 	private async endPlayer(socket_id: string, player_id: number)
 	{
 		// socket.emit('Error', new UnauthorizedException());
-		this.connectedPlayerService.deletePlayer(player_id);
 		this.server.in(socket_id).disconnectSockets(true);
 		this.logger.log(`Ladder Game Server: [ ${player_id} -> ${socket_id} ] disconnected.`);
-	  };
+	};
 
 	async queueProcess()
-	{
-		// [ first-come, first-served basis queue ] //
+	{	try {
 		// if (ladderQueue.length > 1)
 		// {
+		// 	// setGame(ladderQueue[0].id, ladderQueue[1].id);
 		// 	this.setGame(ladderQueue[0].id, ladderQueue[1].id);
 		// 	ladderQueue.splice(0, 2);
 		// }
 
-		// [ ladder basis queue ] //
 		if (ladderQueue.length < 2)
 			return ;
 
 		currentQueueTime = Date.now();
 		if (currentQueueTime - resetQueueTime > 10000) // 1000 milliseconds == 1 second
 		{
-			ladderRange += 500;
+			ladderRange += 100;
 		}
 
 		for (let i = 0; i < ladderQueue.length; ++i)
 		{
 			for (let j = i + 1; j < ladderQueue.length; ++j)
 			{
-				// const player1Ladder = await this.profileService.getLadderById(ladderQueue[i].id);
+				// console.log(ladderQueue[i]);
+				// const ppp = this.connectedPlayerService.getPlayer(ladderQueue[i].id);
+				// console.log(ppp);
+				// const a = await ladderQueue[i].id;
+				// const player1Ladder = await this.connectedPlayerService.getLadderById(a);
 				const player1Ladder = await this.connectedPlayerService.getLadderById(ladderQueue[i].id);
+				// console.log(player1Ladder);
 				const player1less = player1Ladder - ladderRange;
 				const player1greater = player1Ladder + ladderRange;
 				const player2Ladder = await this.connectedPlayerService.getLadderById(ladderQueue[j].id);
@@ -154,31 +214,36 @@ export class LadderGameGateway implements OnGatewayConnection, OnGatewayDisconne
 					ladderQueue.splice(i, 1);
 					ladderQueue.splice(j - 1, 1); // 바로 위에서 요소 하나 삭제되므로 인덱스가 하나씩 당겨짐
 					resetQueueTime = Date.now();
-					ladderRange = 500;
+					ladderRange = 100;
 					return ;
 				}
 			}
 		}
+	} catch (error) {
+		logger2.error(`queueProcess : ${error.message}`);
 	}
-	
-	@WebSocketServer()
-	server: Server;
+	}
 
 	private async setGame(userId1: number, userId2: number)
 	{
-		const currentMatch = await this.matchService.create(userId1, userId2, "ladder");
+		try {
+		const currentMatch = await this.matchService.create(userId1, userId2);
 		if (!currentMatch)
 		{
 			this.endPlayer(await this.connectedPlayerService.getSocketIdById(userId1), userId1);
 			this.endPlayer(await this.connectedPlayerService.getSocketIdById(userId2), userId2);
 			return ;
 		}
-		this.logger.log(`setGame : ${currentMatch} match will soon start!`);
+		this.logger.log(`setGame : match ${currentMatch.match_id} will soon start!`);
 		this.startGame(currentMatch);
+	} catch (error) {
+		logger2.error(`setGame: ${error.message}`);
+	}
 	}
 
 	async startGame(match : MatchEntity)
 	{
+		try {
 		const player1 = await this.connectedPlayerService.getPlayer(match.playerLeft);
 		const player2 = await this.connectedPlayerService.getPlayer(match.playerRight);
 
@@ -203,16 +268,21 @@ export class LadderGameGateway implements OnGatewayConnection, OnGatewayDisconne
 			ballX: 800 / 2,
 			ballY: 600 / 2,
 			ballRadius: 10,
-			ballXvelocity: 3,
-			ballYvelocity: 3,
-			ballSpeed: 3,
+			ballXvelocity: 6,
+			ballYvelocity: 6,
+			ballSpeed: 10,
 			matchId: match.match_id,
 			gameTimer: null,
 		}
 
-		this.logger.log(`startGame : ${match.match_id} -> ${player1.id} vs ${player2.id}`);
+		this.logger.log(`ladder/startGame : ${match.match_id} -> ${player1.id} vs ${player2.id}`);
 		this.gameFieldArr.push(gameField);
-		gameField.gameTimer = setInterval(playGame, 30, this.server, match, player1, player2, gameField);
+		// gameFieldMap.set(match.match_id, gameField);
+		// gameField.gameTimer = setInterval(playGame, 30, this.server, match, player1, player2, gameField);
+		gameField.gameTimer = setInterval(() => {this.playGame(this.server, match, player1, player2, gameField);}, 20);
+	} catch (error) {
+		logger2.error(`startGame : ${error.message}`);
+	}
 	}
 
 	private async getGameFieldByMatchId(match_id: number)
@@ -220,18 +290,27 @@ export class LadderGameGateway implements OnGatewayConnection, OnGatewayDisconne
 		for (let i = 0; i < this.gameFieldArr.length; ++i)
 		{
 			if (this.gameFieldArr[i].matchId === match_id)
-			return this.gameFieldArr[i];
+			{
+				return this.gameFieldArr[i];
+			}
 		}
 	}
 
 	@SubscribeMessage('mouseMove')
 	async movePlayer(@ConnectedSocket() socket: Socket, @MessageBody() userY: number)
 	{
+		try {
 		const player = await this.connectedPlayerService.getPlayerBySocketId(socket.id);
+		if (!player)
+			return ;
 		const match = (await this.matchService.getByPlayerId(player.id));
 		const opponent = await this.matchService.getOpponentByPlayerId(match.match_id, player.id);
 		const gameField = await this.getGameFieldByMatchId(match.match_id);
-		
+
+		// let gameField = gameFieldMap.get(match.match_id);
+
+		// console.log(`match_id is ${match.match_id}, MapSize ${gameFieldMap.size} ,movePlayer : ${gameField}`);
+
 		if (userY < 0)
 		{
 			userY = 0;
@@ -241,18 +320,26 @@ export class LadderGameGateway implements OnGatewayConnection, OnGatewayDisconne
 			userY = 500;
 		}
 
+		const data = {
+			leftY: gameField.paddleLeftY,
+			rightY: gameField.paddleRightY,
+		};
+
 		if (match.playerLeft === player.id)
 		{
 			gameField.paddleLeftY = userY;
-			this.server.to(player.socketId).emit('paddleMove', gameField);
-			this.server.to(opponent.socketId).emit('paddleMove', gameField);
+			this.server.to(player.socketId).emit('paddleMove', data);
+			this.server.to(opponent.socketId).emit('paddleMove', data);
 		}
 		else if (match.playerRight === player.id)
 		{
 			gameField.paddleRightY = userY;
-			this.server.to(player.socketId).emit('paddleMove', gameField);
-			this.server.to(opponent.socketId).emit('paddleMove', gameField);
+			this.server.to(player.socketId).emit('paddleMove', data);
+			this.server.to(opponent.socketId).emit('paddleMove', data);
 		}
+	} catch (error) {
+		logger2.error(`movePlayer : ${error.message}`);
+	}
 	}
 
 	private async updateProfile(winner_id: number, loser_id: number, match_id: number)
@@ -278,10 +365,12 @@ export class LadderGameGateway implements OnGatewayConnection, OnGatewayDisconne
 
 	async endGame(match_id: number, loser_id: number)
 	{
-		this.logger.log(`endGame : match finished.`);
+		try {
+		this.logger.log(`endGame : ${match_id} match finished.`);
 
 		const gameField = await this.getGameFieldByMatchId(match_id);
-		clearInterval(gameField.gameTimer);
+		// let gameField = gameFieldMap.get(match_id);
+		await clearInterval(gameField.gameTimer);
 		const match = await this.matchService.getByMatchId(match_id);
 		let winner_id = 0;
 
@@ -321,51 +410,19 @@ export class LadderGameGateway implements OnGatewayConnection, OnGatewayDisconne
 			this.updateProfile(winner_id, loser_id, match_id);
 		}
 
-		this.sendMatchResult(winner_id, loser_id, true);
-		this.matchService.deleteByMatchId(match.match_id);
+		await this.sendMatchResult(winner_id, loser_id, true);
+		await this.matchService.deleteByMatchId(match.match_id);
 		this.endPlayer((await this.connectedPlayerService.getPlayer(winner_id)).socketId, winner_id);
 		this.endPlayer((await this.connectedPlayerService.getPlayer(loser_id)).socketId, loser_id);
+	} catch (error) {
+			logger2.error(`endGame : ${error.message}`);
+		}
 	}
-}
 
-
-async function collision(b: Ball, p: Paddle)
+	
+	async playGame(server: Server, match: MatchEntity, player1: Player, player2: Player, gameField: GameField)
 {
-	const paddleLocation =
-	{
-	  top: p.y,
-	  bottom: p.y + p.height,
-	  left: p.x,
-	  right: p.x + p.width,
-	};
-
-	const ballLocation =
-	{
-	  top: b.y - b.radius,
-	  bottom: b.y + b.radius,
-	  left: b.x - b.radius,
-	  right: b.x + b.radius,
-	};
-
-	return (
-	  ballLocation.right > paddleLocation.left &&
-	  ballLocation.left < paddleLocation.right &&
-	  ballLocation.top < paddleLocation.bottom &&
-	  ballLocation.bottom > paddleLocation.top
-	);
-  }
-
-async function resetBall(gameField: GameField)
-{
-	gameField.ballX = gameField.canvasWidth / 2;
-	gameField.ballY = gameField.canvasHeight / 2;
-	gameField.ballXvelocity = -3;
-	gameField.ballYvelocity = 3;
-	gameField.ballSpeed = 3;
-}
-
-async function playGame(server: Server, match: MatchEntity, player1: Player, player2: Player, gameField: GameField)
-{
+	try {
 	// location of ball
 	gameField.ballX += gameField.ballXvelocity;
 	gameField.ballY += gameField.ballYvelocity;
@@ -419,26 +476,98 @@ async function playGame(server: Server, match: MatchEntity, player1: Player, pla
 	if (gameField.ballX - gameField.ballRadius < 0)
 	{
 		++gameField.scoreRight;
-		this.MatchService.updateRightScore(match.match_id, gameField.scoreRight);
+		this.matchService.updateRightScore(match.match_id, gameField.scoreRight);
 		
-		if (gameField.scoreRight === 7)
+		if (gameField.scoreRight > 2)
 		{
-			this.endGame(match, null);
+			this.endGame(match.match_id, null);
 		}
-		resetBall(gameField);
+		resetBall(gameField, -1);
 	}
 	else if (gameField.ballX + gameField.ballRadius > gameField.canvasWidth)
 	{
 		++gameField.scoreLeft;
-		this.MatchService.updateLeftScore(match.match_id, gameField.scoreRight);
+		this.matchService.updateLeftScore(match.match_id, gameField.scoreRight);
 
-		if (gameField.scoreLeft === 7)
+		if (gameField.scoreLeft > 2)
 		{
-			this.endGame(match, null);
+			this.endGame(match.match_id, null);
 		}
-		resetBall(gameField);
+		resetBall(gameField, 1);
 	}
 
-	server.to(player1.socketId).emit('updateCanvas', gameField);
-	server.to(player2.socketId).emit('updateCanvas', gameField);
+	
+	//
+	const data = {
+		a: gameField.ballX,
+		b: gameField.ballY,
+		c: gameField.ballXvelocity,
+		d: gameField.ballYvelocity,
+		e: gameField.ballSpeed,
+		f: gameField.scoreLeft,
+		g: gameField.scoreRight,
+	};
+	
+	// server.to(player1.socketId).emit('updateCanvas', gameField);
+	// server.to(player2.socketId).emit('updateCanvas', gameField);
+	server.to(player1.socketId).emit('updateCanvas', data);
+	server.to(player2.socketId).emit('updateCanvas', data);
+
+} catch (error) {
+	logger2.error(`playGame : ${error.message}`);
 }
+}
+
+}
+
+
+async function collision(b: Ball, p: Paddle)
+{
+	const paddleLocation =
+	{
+	  top: p.y,
+	  bottom: p.y + p.height,
+	  left: p.x,
+	  right: p.x + p.width,
+	};
+
+	const ballLocation =
+	{
+	  top: b.y - b.radius,
+	  bottom: b.y + b.radius,
+	  left: b.x - b.radius,
+	  right: b.x + b.radius,
+	};
+
+	return (
+	  ballLocation.right > paddleLocation.left &&
+	  ballLocation.left < paddleLocation.right &&
+	  ballLocation.top < paddleLocation.bottom &&
+	  ballLocation.bottom > paddleLocation.top
+	);
+  }
+
+async function resetBall(gameField: GameField, direction: number)
+{
+	gameField.ballX = gameField.canvasWidth / 2;
+	gameField.ballY = gameField.canvasHeight / 2;
+	gameField.ballXvelocity = 6 * direction;
+	gameField.ballYvelocity = 6;
+	gameField.ballSpeed = 10;
+}
+
+
+
+//
+// async function setGame(userId1: number, userId2: number)
+// {
+// 	const currentMatch = await this.matchService.create(userId1, userId2);
+// 	if (!currentMatch)
+// 	{
+// 		this.endPlayer(await this.connectedPlayerService.getSocketIdById(userId1), userId1);
+// 		this.endPlayer(await this.connectedPlayerService.getSocketIdById(userId2), userId2);
+// 		return ;
+// 	}
+// 	this.logger.log(`setGame : ${currentMatch} match will soon start!`);
+// 	this.startGame(currentMatch);
+// }
